Design exploration of Software categories below features
In this design exploration I put the software categories right below
features so they would be above the fold and more discoverable. There
is an expand function to show all the categories. This is not ideal but
since Gnome has quite a few categories it doesn’t work well to show
them all in the normal state. I also made three small featured tiles
rater than one large one to create a more engaging experience on the
landing screen and give the user more varied exposure to featured
content.